+++ flexiznam/config/default_config.py
@@ -34,7 +34,16 @@ DEFAULT_CONFIG = dict(
         "suite2p_traces",
         "harp",
         "microscopy",
+        "sequencing",
+        "onix",
     ],
     # list of extensions accepted as `MicroscopyData`
     microscopy_extensions=[".czi", ".png", ".gif", ".tif", ".tiff"],
+    # list of extensions accepted as `SequencingData`
+    sequencing_extensions=[".fastq.gz", ".fastq", ".fq.gz", ".fq", ".bam", ".sam"],
+    conda_envs=dict(
+        dlc="dlc_nogui",
+        cottage_analysis="cottage_analysis",
+        suite2p="suite2p",
+    ),
 )